filter.include_exclude_rules: Move filtering rules and related functions here
