+++ augur/filter/io.py
@@ -43,7 +43,7 @@ def filter_kwargs_to_str(kwargs):
 
 
     >>> from augur.dates import numeric_date
-    >>> from augur.filter._run import filter_by_sequence_length, filter_by_date
+    >>> from augur.filter.include_exclude_rules import filter_by_sequence_length, filter_by_date
     >>> sequence_index = pd.DataFrame([{"strain": "strain1", "ACGT": 28000}, {"strain": "strain2", "ACGT": 26000}, {"strain": "strain3", "ACGT": 5000}]).set_index("strain")
     >>> exclude_by = [(filter_by_sequence_length, {"sequence_index": sequence_index, "min_length": 27000})]
     >>> filter_kwargs_to_str(exclude_by[0][1])
